Ebook Chapter 1, part 1 and 2
This book is aimed to help you learn Python from a Biological perspective.

The book has been specially designed for complete beginners and for those coming from a Biological background. 

Whether you're just starting out or looking to expand your skills, this eBook will guide you through the basics of Python with clear explanations and practical examples tailored to the needs of biologists.